+++ pygal/graph/stackedbar.py
@@ -22,9 +22,10 @@ on top of the others instead of being displayed side by side.
 """
 
 from __future__ import division
-from pygal.graph.bar import Bar
-from pygal.util import compute_scale, swap, ident
+
 from pygal.adapters import none_to_zero
+from pygal.graph.bar import Bar
+from pygal.util import ident, swap
 
 
 class StackedBar(Bar):